Refactor signal detail route with modal
