+++ src/app/layout.tsx
@@ -14,8 +14,10 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
+  modal,
 }: {
   children: React.ReactNode;
+  modal: React.ReactNode;
 }) {
   return (
     <html lang="ko" suppressHydrationWarning>
@@ -38,6 +40,7 @@ export default function RootLayout({
             <SignalSearchParamsProvider>
               <Header />
               {children}
+              {modal}
             </SignalSearchParamsProvider>
           </ReactQueryProvider>
         </Suspense>
