+++ src/components/signal/WeeklyActionCountCard.tsx
@@ -5,7 +5,7 @@ import { useWeeklyActionCount } from "@/hooks/useSignal";
 import { GetWeeklyActionCountParams, WeeklyActionCountResponse } from "@/types/signal";
 import { Badge } from "../ui/badge";
 import { cn } from "@/lib/utils";
-import { useSignalSearchParams } from "@/hooks/useSignalSearchParams";
+import { useRouter } from "next/navigation";
 import { CardSkeleton } from "../ui/skeletons";
 
 interface WeeklyActionCountCardProps {
@@ -19,7 +19,7 @@ export const WeeklyActionCountCard: FC<WeeklyActionCountCardProps> = ({
   params,
   data: initialData,
 }) => {
-  const { setParams } = useSignalSearchParams();
+  const router = useRouter();
 
   const { data, isLoading, error } = useWeeklyActionCount(params, {
     select(data) {
@@ -41,7 +41,7 @@ export const WeeklyActionCountCard: FC<WeeklyActionCountCardProps> = ({
   });
 
   const onClickTicker = (ticker: string) => {
-    setParams({ signalId: `${ticker}_OPENAI` });
+    router.push(`/dashboard/d/${ticker}?model=OPENAI`);
   };
 
   if (isLoading) {
